R7 p 29 simulator

+++ src/MachineManagerService/Domain/Domain.Entities/Cargo.cs
@@ -1,9 +1,9 @@
 ﻿namespace R7P.MachineManagerService.Domain.Entities
 {
-    public class Cargo : IEntity<int>
+    public class Cargo : IEntity<long>
     {
-        public int Id { get; set; }
-        public int MachineId { get; set; }
+        public long Id { get; set; }
+        public long MachineId { get; set; }
         public virtual Machine? Machine { get; set; }
         public string? Destination { get; set; }
         public double Volume { get; set; }
